requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-generativeai
requests
youtube-transcript-api
fastapi
uvicorn